move distribution of recipe ratings into callbacks and components
